+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased]
 
+### Added
+
+- Declare `settings.php` predefined variables
+
 ## [0.1.0] - 2025-12-06
 
 ### Added